Change Type for Keys Enum in database, change type player fields in Game

+++ src/SeaBattle/SeaBattle.Lib/Data/Entities/GamePlayer.cs
@@ -12,7 +12,7 @@ namespace SeaBattle.Lib.Entities
     public class GamePlayer : Player, IGamePlayer
     {
         [JsonIgnore]
-        public int PlayerStateId { get; set; }
+        public short PlayerStateId { get; set; }
 
         [ForeignKey(nameof(PlayerStateId))]
         public PlayerState PlayerState { get; set; }
